lesson-07/requirements.txt
Django==1.4.4
South==0.7.6
Whoosh==2.4.1
django-annoying==0.7.7
django-debug-toolbar==0.9.4
django-extensions==1.0.3
django-grappelli==2.4.3
docutils==0.10

http://github.com/toastdriven/django-haystack/tarball/master#egg=django-haystack
